pyproject.toml
[project]
name = "multimodal-analyzer"
version = "0.2.6"
description = "AI-powered media analysis tool for images, audio, and video using multiple LLM providers"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.1.0",
    "pillow>=10.0.0",
    "litellm>=1.72.2",
    "aiofiles>=24.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "tenacity>=8.0.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.2",
    "rich>=13.0.0",
    "nest-asyncio>=1.6.0",
    "ffmpeg-python>=0.2.0",
    "pydub>=0.25.0",
    "pytest-cov>=6.2.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "diskcache>=5.6.3",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "requests>=2.28.0",
    "numpy>=2.3.2",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
multimodal_analyzer_cli = ["prompts/*.md"]

[project.scripts]
multimodal-analyzer = "multimodal_analyzer_cli.cli:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.0.0",
    "types-pillow>=10.0.0",
    "types-pyyaml>=6.0.0",
]

[tool.uv.sources]
multimodal-analyzer = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=multimodal_analyzer_cli --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests requiring API keys",
    "slow: marks tests as slow running",
]
filterwarnings = [
    "ignore::DeprecationWarning:pydub.utils",
    "ignore::DeprecationWarning:pydantic._internal._config",
    "ignore::UserWarning:pydantic.main",
    "ignore::pytest_cov.plugin.CovReportWarning",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",
    "F",
    "I",
    "N",
    "UP",
    "S",
    "B",
    "A",
    "C4",
    "ICN",
    "PIE",
    "T20",
    "Q",
]
ignore = ["S101"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
